% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simstudy.R
\name{generate_simstudy}
\alias{generate_simstudy}
\title{Generate simulated study to test the methods.}
\usage{
generate_simstudy()
}
\description{
Generate simulated transcription (HTseq counts), cnv (copy number segment) and methylation (beta value) data. The study contains metadata for 50 patients (25 controls and 25 cases), 100 genes and 500 methylation probes.
}
